Cleanup and removal of cppyy.gbl
