+++ cmake/FindCppyy.cmake
@@ -95,8 +95,7 @@ function(cppyy_make_python_package _cmpp_target)
     set(_cmpp_file "${_cmpp_file}        cppyy.add_include_path(p)\n")
     set(_cmpp_file "${_cmpp_file}headers = \"${_cmpp_headers}\".split(';')\n")
     set(_cmpp_file "${_cmpp_file}for h in headers:\n")
-    set(_cmpp_file "${_cmpp_file}    inc = os.path.join(\"${_cmpp_PREFIX}\",h)\n")
-    set(_cmpp_file "${_cmpp_file}    cppyy.include(inc)\n")
+    set(_cmpp_file "${_cmpp_file}    cppyy.include(h)\n")
     set(_cmpp_file "${_cmpp_file}dir = os.path.realpath(__file__)\n")
     set(_cmpp_file "${_cmpp_file}dir = os.path.dirname(dir)\n")
     set(_cmpp_file "${_cmpp_file}lib = os.path.join(\"${CMAKE_BINARY_DIR}\", \"${_cmpp_lib}\")\n")
@@ -125,8 +124,9 @@ function(cppyy_make_python_package _cmpp_target)
         set(_cmpp_file "${_cmpp_file}        cppyy.add_include_path(p)\n")
         set(_cmpp_file "${_cmpp_file}headers = \"${_cmpp_headers}\".split(';')\n")
         set(_cmpp_file "${_cmpp_file}for h in headers:\n")
-        set(_cmpp_file "${_cmpp_file}    inc = os.path.join(\"${_cmpp_PREFIX}\",h)\n")
-        set(_cmpp_file "${_cmpp_file}    cppyy.include(inc)\n")
+        set(_cmpp_file "${_cmpp_file}    cppyy.include(h)\n")
+    set(_cmpp_file "${_cmpp_file}from cppyy.gbl import ${_cmpp_NAMESPACE}\n")
+    set(_cmpp_file "${_cmpp_file}from cppyy.gbl import std\n")
         #Write it out
         file(GENERATE OUTPUT ${_cmpp_file_name} CONTENT "${_cmpp_file}")
     endif()
